requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic-settings
python-multipart
tldextract
validators
opencv-python
pyzbar
Pillow
numpy
python-dateutil
python-whois
playwright
sqlalchemy